+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -21,7 +21,7 @@ public class LoginController {
     @RequestMapping(value = "/login", method = RequestMethod.POST)
     public ModelAndView postLogin(HttpServletRequest req, HttpServletResponse res) {
         if (UserSession.isLoggedIn(req.getSession())) {
-            return new ModelAndView(new JspView("redirect:/"));
+            return new ModelAndView(new JspView("redirect:/index.jsp"));
         }
 
         return InMemoryUserRepository.findByAccount(req.getParameter("account"))
@@ -37,7 +37,7 @@ public class LoginController {
         return UserSession.getUserFrom(req.getSession())
                 .map(user -> {
                     log.info("logged in {}", user.getAccount());
-                    return new ModelAndView(new JspView("redirect:/"));
+                    return new ModelAndView(new JspView("redirect:/index.jsp"));
                 })
                 .orElse(new ModelAndView(new JspView("/login.jsp")));
     }
@@ -46,7 +46,7 @@ public class LoginController {
         if (user.checkPassword(request.getParameter("password"))) {
             final HttpSession session = request.getSession();
             session.setAttribute(UserSession.SESSION_KEY, user);
-            return new JspView("redirect:/");
+            return new JspView("redirect:/index.jsp");
         } else {
             return new JspView("redirect:/401.jsp");
         }
