--- app/src/test/java/com/techcourse/ManualHandlerMappingTest.java
@@ -1,35 +1,0 @@
-package com.techcourse;
-
-import jakarta.servlet.http.HttpServletRequest;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.controller.asis.ForwardController;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
-class ManualHandlerMappingTest {
-
-    private ManualHandlerMapping handlerMapping;
-
-    @BeforeEach
-    void setUp() {
-        handlerMapping = new ManualHandlerMapping();
-        handlerMapping.initialize();
-    }
-
-    @Test
-    void getHandler() {
-        final HttpServletRequest request = mock(HttpServletRequest.class);
-
-        when(request.getRequestURI()).thenReturn("/get-test");
-        when(request.getMethod()).thenReturn("GET");
-
-        final Controller controller = handlerMapping.getHandler(request);
-
-        assertThat(controller).usingRecursiveComparison()
-                .isEqualTo(new ForwardController("/get-test.jsp"));
-    }
-}
