+++ mvc/src/main/java/nextstep/mvc/servlet/AnnotationHandlerAdapter.java
@@ -2,7 +2,6 @@ package nextstep.mvc.servlet;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.tobe.HandlerExecution;
 import nextstep.mvc.view.ModelAndView;
