+++ mvc/src/main/java/nextstep/mvc/servlet/AnnotationHandlerMapping.java
@@ -1,8 +1,6 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.servlet;
 
 import jakarta.servlet.http.HttpServletRequest;
-import nextstep.mvc.servlet.ControllerScanner;
-import nextstep.mvc.servlet.HandlerMapping;
 import nextstep.web.annotation.Controller;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.support.RequestMethod;
