+++ mvc/src/main/java/nextstep/mvc/servlet/HandlerKey.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.servlet;
 
 import nextstep.web.support.RequestMethod;
