--- mvc/src/main/java/nextstep/mvc/servlet/ManualHandlerAdapter.java
@@ -1,22 +1,0 @@
-package nextstep.mvc.servlet;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.view.JspView;
-import nextstep.mvc.view.ModelAndView;
-
-public class ManualHandlerAdapter implements HandlerAdapter {
-    @Override
-    public boolean supports(Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        final Controller controller = (Controller) handler;
-        final String viewName = controller.execute(request, response);
-
-        return new ModelAndView(new JspView(viewName));
-    }
-}
