+++ mvc/src/test/java/nextstep/mvc/servlet/AnnotationHandlerMappingTest.java
@@ -1,7 +1,9 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.servlet;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import nextstep.mvc.servlet.AnnotationHandlerMapping;
+import nextstep.mvc.servlet.HandlerExecution;
 import nextstep.mvc.view.ModelAndView;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
